Factor out construction of font-related JSON schemas
